requirements.txt
flask
waitress
openai
python-dotenv
requests
markdown
